+++ bigbluebutton-html5/imports/api/video-streams/server/handlers/userSharedHtml5Webcam.js
