+++ bigbluebutton-html5/imports/api/video-streams/server/handlers/userUnsharedHtml5Webcam.js
