+++ bigbluebutton-html5/imports/api/video-streams/server/methods/userShareWebcam.js
