+++ bigbluebutton-html5/imports/ui/components/subscriptions/component.jsx
@@ -19,7 +19,7 @@ const SUBSCRIPTIONS = [
   'users', 'meetings', 'polls', 'presentations', 'slides', 'slide-positions', 'captions',
   'voiceUsers', 'whiteboard-multi-user', 'screenshare', 'group-chat',
   'presentation-pods', 'users-settings', 'guestUser', 'users-infos', 'note',
-  'network-information', 'ping-pong', 'local-settings', 'users-typing', 'record-meetings', 'video-users',
+  'network-information', 'ping-pong', 'local-settings', 'users-typing', 'record-meetings', 'video-streams',
 ];
 
 class Subscriptions extends Component {
