+++ bigbluebutton-html5/server/main.js
@@ -18,7 +18,7 @@ import '/imports/api/screenshare/server';
 import '/imports/api/users-settings/server';
 import '/imports/api/voice-users/server';
 import '/imports/api/whiteboard-multi-user/server';
-import '/imports/api/video-users/server';
+import '/imports/api/video-streams/server';
 import '/imports/api/network-information/server';
 import '/imports/api/users-infos/server';
 import '/imports/api/note/server';
